Write custom build instructions for igraph
Instead of using igraph's own CMakeLists.txt write one that compilers only a
subset of the source code.

This has the advantage of reducing the external dependencies (both build time
and run time) which makes it easier to build in different environments and
easier to share the prebuilt binaries without needing to find non-standard
external shared libs.

Additionally, before SE2 was "built" as a object library that would need to be
passed as a set of object files to the downstream compilation command in order
to avoid issues with igraph's symbols being in both SE2 and igraph libs. Now
it's easy to compile into a shared or static library that can be linked
against.